refactor(frontend): extract CultureDetail value formatters into cultureDetailFormatters
Move the pure number/distance/seed-rate/package-size formatters, the
persisted-filter shape and the filters storage key out of
cultures/CultureDetail.tsx (1573 -> 1499 lines) into
cultures/cultureDetailFormatters.ts. No behavior changes.
